Complete implementation of find_wall_intersection fns and helpers
- find_closest_horizontal_wall_intersection() computes the unit
coordinates of the closest intersection to a wall
-- it first computes the first intersection with the grid, and then the
offsets required to get to each consecutive horizontal intersection with
the grid.
-- it then checks at each intersection whether a wall exists there or
not (implemented in emit_and_trace_ray()). The ray is constantly traced
until either a wall is reached or we go out of map bounds.
-- if out of bounds return point(INT_MAX, INT_MAX)
-- else return the current intersection x and y unit coords

- find_closest_vertical_wall_intersection() does the same thing except
computes the first_inter_y and inter_offset_y before the x, then calls
emit_and_trace_ray()

- helpers: convert_to_grid_point() and outside_map_bounds()
- convert_to_grid_point() takes a point in unit coords and returns the
grid location where this unit coord lies
- outside_map_bounds() checks whether a given unit_x and unit_y lie
outside of the bounds of the map or not (in our implementation, 4096 x
4096)

- additional: fixed the computation of slice_size in cast_ray() by
performing reverse fishbowl before projecting the distance